requirements.txt
flask>=3.0.0,<4.0
requests>=2.21.0,<3.0
gunicorn
urllib3